+++ src/TestSuite/TestCase.php
@@ -47,13 +47,13 @@ abstract class TestCase extends CakeTestCase
         switch ($name) {
             case 'alias':
                 if (empty($this->_cache['alias'])) {
-                    $this->_cache['alias'] = $this->getAlias($this);
+                    $this->_cache['alias'] = $this->getAlias();
                 }
 
                 return $this->_cache['alias'];
             case 'originClassName':
                 if (empty($this->_cache['originClassName'])) {
-                    $this->_cache['originClassName'] = $this->getOriginClassName($this);
+                    $this->_cache['originClassName'] = $this->getOriginClassName();
                 }
 
                 return $this->_cache['originClassName'];
@@ -94,17 +94,16 @@ abstract class TestCase extends CakeTestCase
      * Gets the alias name for which a test is being performed, starting from a `TestCase` instance.
      *
      * Example: class `MyPlugin\Test\TestCase\Controller\PagesControllerTest`  will return `Pages`.
-     * @param \MeTools\TestSuite\TestCase $class `TestCase` instance
      * @return string The alias name for which a test is being performed
      * @throws \PHPUnit\Framework\AssertionFailedError
      * @throws \ReflectionException
      * @since 2.19.9
      */
-    protected function getAlias(TestCase $class): string
+    protected function getAlias(): string
     {
-        $alias = preg_replace('/^(\w+)(Cell|Controller|Helper|Table|Validator|View)Test$/', '$1', get_class_short_name($class), -1, $count);
+        $alias = preg_replace('/^(\w+)(Cell|Controller|Helper|Table|Validator|View)Test$/', '$1', get_class_short_name($this), -1, $count);
         if (!$alias || !$count) {
-            $this->fail('Unable to get the alias for `' . get_class($class) . '`');
+            $this->fail('Unable to get the alias for `' . get_class($this) . '`');
         }
 
         return $alias;
@@ -114,17 +113,16 @@ abstract class TestCase extends CakeTestCase
      * Gets the class name for which a test is being performed, starting from a `TestCase` class.
      *
      * Example: class `MyPlugin\Test\TestCase\Controller\PagesControllerTest` will return `MyPlugin\Controller\PagesController`.
-     * @param \MeTools\TestSuite\TestCase $className A `TestCase` instance
      * @return class-string The class name for which a test is being performed
      * @since 2.19.2
      * @throw \PHPUnit\Framework\AssertionFailedError
      */
-    protected function getOriginClassName(TestCase $className): string
+    protected function getOriginClassName(): string
     {
-        $originClassName = preg_replace('/^([\w\\\\]+)Test\\\\TestCase\\\\([\w\\\\]+)Test$/', '$1$2', get_class($className), -1, $count);
+        $originClassName = preg_replace('/^([\w\\\\]+)Test\\\\TestCase\\\\([\w\\\\]+)Test$/', '$1$2', get_class($this), -1, $count);
 
         if (!$originClassName || !$count) {
-            $this->fail('Unable to determine the origin class for `' . get_class($className) . '`');
+            $this->fail('Unable to determine the origin class for `' . get_class($this) . '`');
         } elseif (!class_exists($originClassName)) {
             $this->fail('Class `' . $originClassName . '` does not exist');
         }